move the concents tests to a separate file
